fix Non-aiscript error loc & add test
